Move DbLocation, DbRide and DbTaxi to separate namespace. Refactor code to reflect changes.
