+++ TaxiDispatcher.App/Rides/DayInterCityRide.cs
@@ -2,6 +2,7 @@
 using TaxiDispatcher.BL.Locations;
 using TaxiDispatcher.BL.Taxis;
 using TaxiDispatcher.DAL;
+using TaxiDispatcher.DAL.Entities;
 
 namespace TaxiDispatcher.BL.Rides
 {
