+++ TaxiDispatcher.App/Taxis/NaxiTaxi.cs
@@ -1,4 +1,5 @@
 ﻿using TaxiDispatcher.DAL;
+using TaxiDispatcher.DAL.Entities;
 
 namespace TaxiDispatcher.BL.Taxis
 {
