+++ TaxiDispatcher.App/Taxis/Taxi.cs
@@ -3,6 +3,7 @@ using System.Linq;
 using TaxiDispatcher.BL.Locations;
 using TaxiDispatcher.BL.Rides;
 using TaxiDispatcher.DAL;
+using TaxiDispatcher.DAL.Entities;
 
 namespace TaxiDispatcher.BL.Taxis
 {
