+++ TaxiDispatcher.Client.Tests/HelperClasses/TestDatabase.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
 using TaxiDispatcher.DAL;
+using TaxiDispatcher.DAL.Entities;
 
 namespace TaxiDispatcher.Tests.HelperClasses
 {
