+++ TaxiDispatcher.DAL/InMemoryDataBase.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using TaxiDispatcher.DAL.Entities;
 
 namespace TaxiDispatcher.DAL
 {
