+++ TaxiDispatcher.DAL/TaxiDispatcher.DAL.csproj
@@ -41,12 +41,12 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="DbLocation.cs" />
-    <Compile Include="DbTaxi.cs" />
+    <Compile Include="Entities\DbLocation.cs" />
+    <Compile Include="Entities\DbTaxi.cs" />
     <Compile Include="IDatabase.cs" />
     <Compile Include="InMemoryDataBase.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="DbRide.cs" />
+    <Compile Include="Entities\DbRide.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
